+++ tests/test_digital.py
@@ -9,31 +9,15 @@ import torch
 from torch import Tensor
 
 import pyqtorch as pyq
-from pyqtorch.apply import apply_density_mat, apply_operator, operator_product
+from pyqtorch.apply import apply_operator
 from pyqtorch.matrices import (
     DEFAULT_MATRIX_DTYPE,
-    HMAT,
-    IMAT,
-    XMAT,
-    YMAT,
-    ZMAT,
-    _dagger,
-)
-from pyqtorch.noise import (
-    AmplitudeDamping,
-    GeneralizedAmplitudeDamping,
-    Noise,
-    PhaseDamping,
 )
 from pyqtorch.parametric import Parametric
-from pyqtorch.primitive import H, I, Primitive, X, Y, Z
+from pyqtorch.primitive import Primitive
 from pyqtorch.utils import (
     ATOL,
-    DensityMatrix,
-    density_mat,
-    operator_kron,
     product_state,
-    random_state,
 )
 
 state_000 = product_state("000")
@@ -346,174 +330,6 @@ def test_U() -> None:
     assert u.spectral_gap == 2.0
 
 
-def test_dm(n_qubits: int, batch_size: int) -> None:
-    state = random_state(n_qubits)
-    projector = torch.outer(state.flatten(), state.conj().flatten()).view(
-        2**n_qubits, 2**n_qubits, 1
-    )
-    dm = density_mat(state)
-    assert dm.size() == torch.Size([2**n_qubits, 2**n_qubits, 1])
-    assert torch.allclose(dm, projector)
-    assert torch.allclose(dm.squeeze(), dm.squeeze() @ dm.squeeze())
-    states = []
-    projectors = []
-    for batch in range(batch_size):
-        state = random_state(n_qubits)
-        states.append(state)
-        projector = torch.outer(state.flatten(), state.conj().flatten()).view(
-            2**n_qubits, 2**n_qubits, 1
-        )
-        projectors.append(projector)
-    dm_proj = torch.cat(projectors, dim=2)
-    state_cat = torch.cat(states, dim=n_qubits)
-    dm = density_mat(state_cat)
-    assert dm.size() == torch.Size([2**n_qubits, 2**n_qubits, batch_size])
-    assert torch.allclose(dm, dm_proj)
-
-
-@pytest.mark.parametrize("n_qubits", [{"low": 2, "high": 5}], indirect=True)
-def test_operator_product(
-    random_unitary_gate: Primitive | Parametric,
-    n_qubits: int,
-) -> None:
-    batch_size_1 = torch.randint(low=1, high=5, size=(1,)).item()
-    batch_size_2 = torch.randint(low=1, high=5, size=(1,)).item()
-    max_batch = max(batch_size_2, batch_size_1)
-    values = {"theta": torch.rand(1)}
-    op = random_unitary_gate.tensor(values=values, full_support=tuple(range(n_qubits)))
-    op_mul = operator_product(
-        op.repeat(1, 1, batch_size_1), _dagger(op.repeat(1, 1, batch_size_2))
-    )
-    assert op_mul.size() == torch.Size([2**n_qubits, 2**n_qubits, max_batch])
-    assert torch.allclose(
-        op_mul,
-        torch.eye(2**n_qubits, dtype=torch.cdouble)
-        .unsqueeze(2)
-        .repeat(1, 1, max_batch),
-    )
-
-
-@pytest.mark.parametrize("n_qubits", [{"low": 2, "high": 5}], indirect=True)
-def test_apply_density_mat(
-    random_unitary_gate: Primitive | Parametric,
-    n_qubits: int,
-    batch_size: int,
-    random_input_dm: DensityMatrix,
-) -> None:
-    values = {"theta": torch.rand(1)}
-    op = random_unitary_gate.tensor(values=values, full_support=tuple(range(n_qubits)))
-    rho = random_input_dm
-    rho_evol = apply_density_mat(op, rho)
-    assert rho_evol.size() == torch.Size([2**n_qubits, 2**n_qubits, batch_size])
-    rho_expected = operator_product(op, operator_product(rho, _dagger(op)))
-    assert torch.allclose(rho_evol, rho_expected)
-
-
-@pytest.mark.parametrize(
-    "operator,matrix", [(I, IMAT), (X, XMAT), (Z, ZMAT), (Y, YMAT), (H, HMAT)]
-)
-def test_operator_kron(operator: Tensor, matrix: Tensor) -> None:
-    n_qubits = torch.randint(low=1, high=5, size=(1,)).item()
-    batch_size = torch.randint(low=1, high=5, size=(1,)).item()
-    states, krons = [], []
-    for batch in range(batch_size):
-        state = random_state(n_qubits)
-        states.append(state)
-        kron = torch.kron(density_mat(state).squeeze(), matrix).unsqueeze(2)
-        krons.append(kron)
-    input_state = torch.cat(states, dim=n_qubits)
-    kron_out = operator_kron(density_mat(input_state), operator(0).dagger())
-    assert kron_out.size() == torch.Size(
-        [2 ** (n_qubits + 1), 2 ** (n_qubits + 1), batch_size]
-    )
-    kron_expect = torch.cat(krons, dim=2)
-    assert torch.allclose(kron_out, kron_expect)
-    assert torch.allclose(
-        torch.kron(operator(0).dagger().contiguous(), I(0).unitary()),
-        operator_kron(operator(0).dagger(), I(0).unitary()),
-    )
-
-
-def test_kron_batch() -> None:
-    n_qubits = torch.randint(low=1, high=5, size=(1,)).item()
-    batch_size_1 = torch.randint(low=1, high=5, size=(1,)).item()
-    batch_size_2 = torch.randint(low=1, high=5, size=(1,)).item()
-    max_batch = max(batch_size_1, batch_size_2)
-    dm_1 = density_mat(random_state(n_qubits, batch_size_1))
-    dm_2 = density_mat(random_state(n_qubits, batch_size_2))
-    dm_out = operator_kron(dm_1, dm_2)
-    assert dm_out.size() == torch.Size(
-        [2 ** (2 * n_qubits), 2 ** (2 * n_qubits), max_batch]
-    )
-    if batch_size_1 > batch_size_2:
-        dm_2 = dm_2.repeat(1, 1, batch_size_1)[:, :, :batch_size_1]
-    elif batch_size_2 > batch_size_1:
-        dm_1 = dm_1.repeat(1, 1, batch_size_2)[:, :, :batch_size_2]
-    density_matrices = []
-    for batch in range(max_batch):
-        density_matrice = torch.kron(dm_1[:, :, batch], dm_2[:, :, batch]).unsqueeze(2)
-        density_matrices.append(density_matrice)
-    dm_expect = torch.cat(density_matrices, dim=2)
-    assert torch.allclose(dm_out, dm_expect)
-
-
-@pytest.mark.parametrize("n_qubits", [{"low": 2, "high": 5}], indirect=True)
-def test_flip_gates(
-    n_qubits: int,
-    target: int,
-    batch_size: int,
-    rho_input: Tensor,
-    random_flip_gate: Noise,
-    flip_expected_state: DensityMatrix,
-    flip_probability: Tensor | float,
-    flip_gates_prob_0: Noise,
-    flip_gates_prob_1: tuple,
-    random_input_dm: DensityMatrix,
-) -> None:
-    FlipGate = random_flip_gate
-    output_state: DensityMatrix = FlipGate(target, flip_probability)(rho_input)
-    assert output_state.size() == torch.Size([2**n_qubits, 2**n_qubits, batch_size])
-    assert torch.allclose(output_state, flip_expected_state)
-
-    input_state = random_input_dm  # fix the same random state for every call
-    assert torch.allclose(flip_gates_prob_0(input_state), input_state)
-
-    FlipGate_1, expected_op = flip_gates_prob_1
-    assert torch.allclose(FlipGate_1(input_state), expected_op)
-
-
-def test_damping_gates(
-    n_qubits: int,
-    target: int,
-    batch_size: int,
-    damping_expected_state: tuple,
-    damping_gates_prob_0: Tensor,
-    random_input_dm: DensityMatrix,
-    rho_input: Tensor,
-) -> None:
-    DampingGate, expected_state = damping_expected_state
-    apply_gate = DampingGate(rho_input)
-    assert apply_gate.size() == torch.Size([2**n_qubits, 2**n_qubits, batch_size])
-    assert torch.allclose(apply_gate, expected_state)
-
-    input_state = random_input_dm
-    assert torch.allclose(
-        damping_gates_prob_0(input_state),
-        I(target)(input_state),
-    )
-
-    rho_0: DensityMatrix = density_mat(product_state("0", batch_size))
-    rho_1: DensityMatrix = density_mat(product_state("1", batch_size))
-    if DampingGate == AmplitudeDamping:
-        assert torch.allclose(DampingGate(target, rate=1)(rho_1), rho_0)
-    elif DampingGate == PhaseDamping:
-        assert torch.allclose(DampingGate(target, rate=1)(rho_1), I(target)(rho_1))
-    elif DampingGate == GeneralizedAmplitudeDamping:
-        assert torch.allclose(
-            DampingGate(target, error_probability=(1, 1))(rho_1), rho_0
-        )
-
-
 @pytest.mark.parametrize("gate", [pyq.RX, pyq.RY, pyq.RZ])
 def test_parametric_constantparam(gate: pyq.parametric.Parametric) -> None:
     n_qubits = 4
@@ -525,21 +341,3 @@ def test_parametric_constantparam(gate: pyq.parametric.Parametric) -> None:
         gate(target, "theta")(state, {"theta": param_val}),
         gate(target, param_val)(state),
     )
-
-
-@pytest.mark.parametrize("n_qubits", [{"low": 2, "high": 5}], indirect=True)
-def test_noisy_primitive(
-    random_noisy_unitary_gate: tuple,
-    random_input_dm: DensityMatrix,
-    n_qubits: int,
-    batch_size: int,
-) -> None:
-    noisy_primitive, primitve_gate, noise_gate = random_noisy_unitary_gate
-    breakpoint()
-    state = random_input_dm
-    values = {"theta": torch.rand(1)}
-    rho_evol = noisy_primitive(state, values)
-    assert rho_evol.size() == torch.Size([2**n_qubits, 2**n_qubits, batch_size])
-    rho_expected = noise_gate(primitve_gate(state, values))
-    assert rho_expected.size() == torch.Size([2**n_qubits, 2**n_qubits, batch_size])
-    assert torch.allclose(rho_evol, rho_expected)
